Instantaneously show background after changing points
