+++ web_app/src/components/ImageOverlay.vue
@@ -5,6 +5,8 @@ import BoundingBox from "./BoundingBox.vue";
 import SelectionPoint from "./SelectionPoint.vue";
 import { ref, onMounted, computed } from "vue";
 import LoadingSpinner from "./LoadingSpinner.vue";
+import { config, endpoints } from "@/config";
+import { createMaskImage, sendRequest } from "@/utils";
 
 
 const imageState = useImageStateStore();
@@ -72,12 +74,50 @@ function scaleOverlay() {
     }
 }
 
+function sendPoints() {
+    if (imageState.points.length === 0) {
+        viewState.showBackground = false;
+        return;
+    }
+
+    const requestUri = config.serverUri + endpoints.sendSelection.replace("{image_id}", imageState.imageId.toString());
+    const requestData = JSON.stringify({"data": imageState.points});
+    const responsePromise = sendRequest(requestUri, requestData, "PUT");
+
+    viewState.isWaitingForResponse = true;
+
+    // Backend returns a background mask
+    responsePromise.then((response) => {
+        viewState.isWaitingForResponse = false;
+
+        const maskImageData = createMaskImage(response.data.mask);
+        const canvas = document.createElement("canvas");
+        const ctx = canvas.getContext("2d");
+        if (ctx == undefined) return;
+
+        viewState.showBackground = true;
+
+        ctx.canvas.width = imageState.width;
+        ctx.canvas.height = imageState.height;
+        ctx.putImageData(maskImageData, 0, 0);
+
+        const maskImage = new Image();
+        maskImage.onload = () => {
+            ctx.drawImage(maskImage, 0, 0);
+        };
+
+        document.querySelector<HTMLImageElement>("#mask-image")!.src = canvas.toDataURL();
+    });
+}
+
 function handleOverlayClick(event: MouseEvent) {
     if (imageState.isPanning) return;
 
     const bbox = (event.target! as HTMLElement).getBoundingClientRect();
+    // [x, y] relative to original image size, not scaled
     const x = (event.clientX - bbox.left) / scale.value / imageState.userZoom;
     const y = (event.clientY - bbox.top) / scale.value / imageState.userZoom;
+    const beforePointCount = imageState.points.length;
 
     if (viewState.isAddingPoint) {
         if ((event.target! as HTMLElement).classList.contains("selection-point")) return;
@@ -91,8 +131,13 @@ function handleOverlayClick(event: MouseEvent) {
             const pointY = Number((event.target! as HTMLElement).getAttribute("data-y"));
             imageState.removeNearbyPoint(pointX, pointY);
         }
-        else imageState.removeNearbyPoint(x, y);
+        else {
+            imageState.removeNearbyPoint(x, y);
+        }
     }
+
+    // If point was added or removed, send the updated points to the server
+    if (beforePointCount !== imageState.points.length) sendPoints();
 }
